pyproject.toml
[tool.poetry]
name = "django-gdparchis"
version = "0.1.0"
description = "Backend for gdparchis"
authors = ["turulomio <turulomio@yahoo.es>"]
license = "GPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
request_casting = "^0.7.0"
pydicts = "^0.16.0"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
